python/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "araali"
version = "0.0.3"
authors = [
  { name="Abhishek R. Singh", email="abhishek@araalinetworks.com" },
]
description = "Araali API"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "grpcio",
  "protobuf",
]

[project.urls]
"Homepage" = "https://github.com/araalinetworks/api/tree/main/python/api"
"Bug Tracker" = "https://github.com/araalinetworks/api/issues"

[tool]
